+++ src/main/java/nextstep/web/common/controller/Response.java
@@ -1,11 +1,12 @@
-package nextstep.web.exception;
+package nextstep.web.common.controller;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public class ErrorResponse {
+public class Response<T> {
     private final int status;
     private final String message;
+    private final T data;
 }
