+++ src/main/java/nextstep/web/common/exception/CommonErrorCode.java
@@ -1,4 +1,4 @@
-package nextstep.web.exception;
+package nextstep.web.common.exception;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
