+++ src/main/java/nextstep/web/common/exception/ErrorResponse.java
@@ -1,12 +1,11 @@
-package nextstep.web.controller;
+package nextstep.web.common.exception;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public class Response<T> {
+public class ErrorResponse {
     private final int status;
     private final String message;
-    private final T data;
 }
