+++ src/main/java/nextstep/web/common/repository/RoomEscapeRepository.java
@@ -1,4 +1,4 @@
-package nextstep.web.repository;
+package nextstep.web.common.repository;
 
 public interface RoomEscapeRepository<T> {
